+++ lib/legion/extensions/node/transport/queues/cluster_control.rb
@@ -1,0 +1,22 @@
+# frozen_string_literal: true
+
+module Legion
+  module Extensions
+    module Node
+      module Transport
+        module Queues
+          class ClusterControl < Legion::Transport::Queue
+            def queue_name
+              "legion.cluster.control.#{Legion::Settings[:client][:name]}"
+            end
+
+            def queue_options
+              { durable: false, exclusive: false, auto_delete: true,
+                arguments: { 'x-queue-type': 'classic' } }
+            end
+          end
+        end
+      end
+    end
+  end
+end